Refactored status conditions to enable better synchronization between controllers
