+++ config/samples/mariadb_v1alpha1_mariadb_from_backup.yaml
@@ -30,7 +30,7 @@ spec:
 
   bootstrapFrom:
     backupRef:
-      name: backup-scheduled
+      name: backup
 
   resources:
     requests:
