+++ pkg/conditions/conditions.go
@@ -1,211 +1,134 @@
 package conditions
 
 import (
-	"errors"
+	"context"
 	"fmt"
+	"reflect"
 
-	mariadbv1alpha1 "github.com/mariadb-operator/mariadb-operator/api/v1alpha1"
-	"github.com/mariadb-operator/mariadb-operator/pkg/statefulset"
 	batchv1 "k8s.io/api/batch/v1"
-	corev1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 type Conditioner interface {
 	SetCondition(condition metav1.Condition)
 }
 
-func SetReadyHealthty(c Conditioner) {
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeReady,
-		Status:  metav1.ConditionTrue,
-		Reason:  mariadbv1alpha1.ConditionReasonHealthy,
-		Message: "Healthy",
-	})
+type Patcher func(Conditioner)
+
+type Ready struct{}
+
+func NewReady() *Ready {
+	return &Ready{}
 }
 
-func SetReadyUnhealthtyWithError(c Conditioner, err error) {
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeReady,
-		Status:  metav1.ConditionFalse,
-		Reason:  mariadbv1alpha1.ConditionReasonHealthy,
-		Message: err.Error(),
-	})
-}
-
-func SetReadyCreatedWithMessage(c Conditioner, message string) {
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeReady,
-		Status:  metav1.ConditionTrue,
-		Reason:  mariadbv1alpha1.ConditionReasonCreated,
-		Message: message,
-	})
-}
-
-func SetReadyCreated(c Conditioner) {
-	SetReadyCreatedWithMessage(c, "Created")
-}
-
-func SetReadyFailedWithMessage(c Conditioner, message string) {
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeReady,
-		Status:  metav1.ConditionFalse,
-		Reason:  mariadbv1alpha1.ConditionReasonFailed,
-		Message: message,
-	})
-}
-
-func SetReadyFailed(c Conditioner) {
-	SetReadyFailedWithMessage(c, "Failed")
-}
-
-func SetConfiguringReplication(c Conditioner, mariadb *mariadbv1alpha1.MariaDB) error {
-	if mariadb.Spec.Replication == nil {
-		return errors.New("replication required to set configuring replication Ready condition")
+func (p *Ready) PatcherFailed(msg string) Patcher {
+	return func(c Conditioner) {
+		SetReadyFailedWithMessage(c, msg)
 	}
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeReady,
-		Status:  metav1.ConditionFalse,
-		Reason:  mariadbv1alpha1.ConditionReasonConfigureReplication,
-		Message: "Configuring replication",
-	})
-	return nil
 }
 
-func SetPrimarySwitching(c Conditioner, mariadb *mariadbv1alpha1.MariaDB) error {
-	if mariadb.Spec.Replication == nil {
-		return errors.New("replication required to set switching primary Ready condition")
-	}
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeReady,
-		Status:  metav1.ConditionFalse,
-		Reason:  mariadbv1alpha1.ConditionReasonSwitchPrimary,
-		Message: switchingPrimaryMessage(mariadb),
-	})
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypePrimarySwitched,
-		Status:  metav1.ConditionFalse,
-		Reason:  mariadbv1alpha1.ConditionReasonSwitchoverInProgress,
-		Message: switchingPrimaryMessage(mariadb),
-	})
-	return nil
-}
-
-func SetPrimarySwitched(c Conditioner) {
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypePrimarySwitched,
-		Status:  metav1.ConditionTrue,
-		Reason:  mariadbv1alpha1.ConditionReasonSwitchoverComplete,
-		Message: "Switchover complete",
-	})
-}
-
-func SetCompleteWithCronJob(c Conditioner, cronJob *batchv1.CronJob) {
-	setScheduled := func() {
-		c.SetCondition(metav1.Condition{
-			Type:    mariadbv1alpha1.ConditionTypeComplete,
-			Status:  metav1.ConditionFalse,
-			Reason:  mariadbv1alpha1.ConditionReasonCronJobScheduled,
-			Message: "Scheduled",
-		})
-	}
-
-	if cronJob.Status.LastScheduleTime == nil || cronJob.Status.LastSuccessfulTime == nil {
-		setScheduled()
-		return
-	}
-	if cronJob.Status.LastSuccessfulTime.Before(cronJob.Status.LastScheduleTime) {
-		if len(cronJob.Status.Active) > 0 {
-			c.SetCondition(metav1.Condition{
-				Type:    mariadbv1alpha1.ConditionTypeComplete,
-				Status:  metav1.ConditionFalse,
-				Reason:  mariadbv1alpha1.ConditionReasonCronJobRunning,
-				Message: "Running",
-			})
+func (p *Ready) PatcherWithError(err error) Patcher {
+	return func(c Conditioner) {
+		if err == nil {
+			SetReadyCreated(c)
 		} else {
-			c.SetCondition(metav1.Condition{
-				Type:    mariadbv1alpha1.ConditionTypeComplete,
-				Status:  metav1.ConditionFalse,
-				Reason:  mariadbv1alpha1.ConditionReasonCronJobFailed,
-				Message: "Failed",
-			})
+			SetReadyFailed(c)
 		}
-		return
-	}
-
-	if cronJob.Status.LastScheduleTime.Equal(cronJob.Status.LastSuccessfulTime) ||
-		cronJob.Status.LastScheduleTime.Before(cronJob.Status.LastSuccessfulTime) {
-		c.SetCondition(metav1.Condition{
-			Type:    mariadbv1alpha1.ConditionTypeComplete,
-			Status:  metav1.ConditionTrue,
-			Reason:  mariadbv1alpha1.ConditionReasonCronJobSuccess,
-			Message: "Success",
-		})
-		return
-	}
-
-	setScheduled()
-}
-
-func SetCompleteWithJob(c Conditioner, job *batchv1.Job) {
-	switch getJobConditionType(job) {
-	case batchv1.JobFailed:
-		c.SetCondition(metav1.Condition{
-			Type:    mariadbv1alpha1.ConditionTypeComplete,
-			Status:  metav1.ConditionTrue,
-			Reason:  mariadbv1alpha1.ConditionReasonJobFailed,
-			Message: "Failed",
-		})
-	case batchv1.JobComplete:
-		c.SetCondition(metav1.Condition{
-			Type:    mariadbv1alpha1.ConditionTypeComplete,
-			Status:  metav1.ConditionTrue,
-			Reason:  mariadbv1alpha1.ConditionReasonJobComplete,
-			Message: "Success",
-		})
-	case batchv1.JobSuspended:
-		c.SetCondition(metav1.Condition{
-			Type:    mariadbv1alpha1.ConditionTypeComplete,
-			Status:  metav1.ConditionFalse,
-			Reason:  mariadbv1alpha1.ConditionReasonJobSuspended,
-			Message: "Suspended",
-		})
-	default:
-		c.SetCondition(metav1.Condition{
-			Type:    mariadbv1alpha1.ConditionTypeComplete,
-			Status:  metav1.ConditionFalse,
-			Reason:  mariadbv1alpha1.ConditionReasonJobRunning,
-			Message: "Running",
-		})
 	}
 }
 
-func SetCompleteFailedWithMessage(c Conditioner, message string) {
-	c.SetCondition(metav1.Condition{
-		Type:    mariadbv1alpha1.ConditionTypeComplete,
-		Status:  metav1.ConditionFalse,
-		Reason:  mariadbv1alpha1.ConditionReasonFailed,
-		Message: message,
-	})
-}
-
-func SetCompleteFailed(c Conditioner) {
-	SetCompleteFailedWithMessage(c, "Failed")
-}
-
-func getJobConditionType(job *batchv1.Job) batchv1.JobConditionType {
-	for _, c := range job.Status.Conditions {
-		if c.Status == corev1.ConditionFalse {
-			continue
+func (p *Ready) PatcherRefResolver(err error, obj interface{}) Patcher {
+	return func(c Conditioner) {
+		if err == nil {
+			return
 		}
-		return c.Type
+		if apierrors.IsNotFound(err) {
+			SetReadyFailedWithMessage(c, fmt.Sprintf("%s not found", getType(obj)))
+			return
+		}
+		SetReadyFailedWithMessage(c, fmt.Sprintf("Error getting %s", getType(obj)))
 	}
-	return ""
 }
 
-func switchingPrimaryMessage(mariadb *mariadbv1alpha1.MariaDB) string {
-	return fmt.Sprintf(
-		"Switching primary to '%s'",
-		statefulset.PodName(mariadb.ObjectMeta, mariadb.Spec.Replication.Primary.PodIndex),
-	)
+func (p *Ready) PatcherHealthy(err error) Patcher {
+	return func(c Conditioner) {
+		if err == nil {
+			SetReadyHealthty(c)
+		} else {
+			SetReadyUnhealthtyWithError(c, err)
+		}
+	}
+}
+
+type Complete struct {
+	client client.Client
+}
+
+func NewComplete(client client.Client) *Complete {
+	return &Complete{
+		client: client,
+	}
+}
+
+func (p *Complete) PatcherFailed(msg string) Patcher {
+	return func(c Conditioner) {
+		SetCompleteFailedWithMessage(c, msg)
+	}
+}
+
+func (p *Complete) PatcherWithCronJob(ctx context.Context, err error, key types.NamespacedName) (Patcher, error) {
+	if err != nil {
+		return func(c Conditioner) {
+			SetCompleteFailedWithMessage(c, "Error creating CronJob")
+		}, nil
+	}
+
+	var cronJob batchv1.CronJob
+	if err := p.client.Get(ctx, key, &cronJob); err != nil {
+		return nil, err
+	}
+	return func(c Conditioner) {
+		SetCompleteWithCronJob(c, &cronJob)
+	}, nil
+}
+
+func (p *Complete) PatcherWithJob(ctx context.Context, err error, key types.NamespacedName) (Patcher, error) {
+	if err != nil {
+		return func(c Conditioner) {
+			SetCompleteFailedWithMessage(c, "Error creating Job")
+		}, nil
+	}
+
+	var job batchv1.Job
+	if err := p.client.Get(ctx, key, &job); err != nil {
+		return nil, err
+	}
+	return func(c Conditioner) {
+		SetCompleteWithJob(c, &job)
+	}, nil
+}
+
+func (p *Complete) PatcherRefResolver(err error, obj runtime.Object) Patcher {
+	return func(c Conditioner) {
+		if err == nil {
+			return
+		}
+		if apierrors.IsNotFound(err) {
+			SetCompleteFailedWithMessage(c, fmt.Sprintf("%s not found", getType(obj)))
+			return
+		}
+		SetCompleteFailedWithMessage(c, fmt.Sprintf("Error getting %s", getType(obj)))
+	}
+}
+
+func getType(obj interface{}) string {
+	if t := reflect.TypeOf(obj); t.Kind() == reflect.Ptr {
+		return t.Elem().Name()
+	} else {
+		return t.Name()
+	}
 }
